+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 > And empty the **[Unreleased]** section above.
 ----
 
+## Added
+
+- 📖 Documentation on how to use custom templates
+
 ## [0.8.6] - 2022-11-09
 
 ### Changed